+++ ChangeLog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ## Version 2.9 - 14/06/2021
 
+- FIX : Change default rights to 0 *01/07/2021* - 2.9.3
 - FIX : Compatibility V13 *17/05/2021* - 2.9.2
 - FIX - Compatibility V14 : Edit the descriptor: family - *2021-06-10* - 2.9.1
 - NEW : TK2003-0572 - Qualiopi Référents Ajouter les tags DocEdit qui permettent d'y avoir acces *04/06/2021* - 2.9